Refactor out record long tasks into a separate function
