+++ packages/opencensus-web-instrumentation-perf/src/initial-load-root-span.ts
@@ -15,7 +15,7 @@
  */
 
 import {Annotation, ATTRIBUTE_HTTP_URL, ATTRIBUTE_HTTP_USER_AGENT, ATTRIBUTE_LONG_TASK_ATTRIBUTION, ATTRIBUTE_NAV_TYPE, parseUrl, RootSpan, Span, SpanKind, Tracer} from '@opencensus/web-core';
-import {GroupedPerfEntries} from './perf-recorder';
+import {GroupedPerfEntries} from './perf-grouper';
 import {PerformanceLongTaskTiming, PerformanceNavigationTimingExtended} from './perf-types';
 import {getResourceSpan} from './resource-span';
 import {annotationsForPerfTimeFields} from './util';
